Allow for amx to use fuzzy match when using Helm backend.
This commit changes the current behaviour and it is waiting for revision.

One problem that I see is that, although fuzzy matching is more
convenient, not all users use it. In this case there should be an user
option for turning on/off. Because the backend set up in neat way I
found kind of awkward to write such option in there, but in principle
we can fetch for `helm-completion-styles` to see if `helm-fuzzy` is
enabled or for `emacs` and see if `completion-style' has `flex` or
`helm-flex`.